src/python/requirements.txt
pyusb
cherrypy
